Control.Concurrent.CVar.Strict

Description

Strict alternatives to the functions in Control.Monad.Conc.CVar. Specifically, values are evaluated to normal form befire being put into a CVar.

CVars

newEmptyCVar :: (MonadConc m, NFData a) => m (CVar m a)

Create a new empty CVar.

newCVar :: (MonadConc m, NFData a) => a -> m (CVar m a)

Create a new CVar containing a value.

takeCVar :: MonadConc m => CVar m a -> m a

Take a value from a CVar. This "empties" the CVar, allowing a new value to be put in. This will block if there is no value in the CVar already, until one has been put.

takeCVar cvar = tryTakeCVar cvar >>= maybe (takeCVar cvar) return

putCVar :: (MonadConc m, NFData a) => CVar m a -> a -> m ()

Put a value into a CVar. If there is already a value there, this will block until that value has been taken, at which point the value will be stored.

readCVar :: MonadConc m => CVar m a -> m a

Block until a value is present in the CVar, and then return it. As with readMVar, this does not "remove" the value, multiple reads are possible.

swapCVar :: (MonadConc m, NFData a) => CVar m a -> a -> m a

Swap the contents of a CVar, and return the value taken.

tryTakeCVar :: MonadConc m => CVar m a -> m (Maybe a)

Attempt to take a value from a CVar, returning a Just (and emptying the CVar) if there was something there, otherwise returning Nothing.

tryPutCVar :: (MonadConc m, NFData a) => CVar m a -> a -> m Bool

Attempt to put a value in a CVar, returning True (and filling the CVar) if there was nothing there, otherwise returning False.

isEmptyCVar :: MonadConc m => CVar m a -> m Bool

Check if a CVar is empty.

withCVar :: MonadConc m => CVar m a -> (a -> m b) -> m b

Operate on the contents of a CVar, replacing the contents after finishing.

modifyCVar_ :: (MonadConc m, NFData a) => CVar m a -> (a -> m a) -> m ()

Modify the contents of a CVar.

modifyCVar :: (MonadConc m, NFData a) => CVar m a -> (a -> m (a, b)) -> m b

Apply a function to the value inside a CVar, and also return a value.

Binary semaphores

A common use of CVars is in making binary semaphores to control mutual exclusion over a resource, so a couple of helper functions are provided.

lock :: MonadConc m => CVar m () -> m ()

Put a () into a CVar, claiming the lock. This is atomic.

unlock :: MonadConc m => CVar m () -> m ()

Empty a CVar, releasing the lock. This is atomic.