Control.Monad.Conc.CVar

Description

Combinators using CVars. These provide many of the helpful functions found in Control.Concurrent.MVar, but for CVars. Note that these do not in general mask exceptions, and are not atomic, being implemented in terms of the primitives in the ConcFuture and ConcCVar typeclasses.

Combinators

newCVar :: ConcCVar cvar m => a -> m (cvar a)

Create a new CVar containing a value.

swapCVar :: ConcCVar cvar m => cvar a -> a -> m a

Swap the contents of a CVar, and return the value taken.

isEmptyCVar :: ConcCVar cvar m => cvar a -> m Bool

Check if a CVar is empty.

withCVar :: ConcCVar cvar m => cvar a -> (a -> m b) -> m b

Operate on the contents of a CVar, replacing the contents after finishing.

modifyCVar :: ConcCVar cvar m => cvar a -> (a -> m (a, b)) -> m b

Apply a function to the value inside a CVar, and also return a value.

modifyCVar_ :: ConcCVar cvar m => cvar a -> (a -> m a) -> m ()

Modify the contents of a CVar.

Binary semaphores

A common use of CVars is in making binary semaphores to control mutual exclusion over a resource, so a couple of helper functions are provided.

lock :: ConcCVar cvar m => cvar () -> m ()

Put a () into a CVar, claiming the lock. This is atomic.

unlock :: ConcCVar cvar m => cvar () -> m ()

Empty a CVar, releasing the lock. This is atomic.