Control.Monad.Conc.Fixed

Description

Concurrent monads with a fixed scheduler.

The Conc Monad

data Conc t a

The Conc monad itself. Under the hood, this uses continuations so it's able to interrupt and resume a monadic computation at any point where a primitive is used.

This uses the same universally-quantified indexing state trick as used by ST and STRefs to prevent mutable references from leaking out of the monad. See runConc for an example of what this means.

runConc :: Scheduler s -> s -> (forall t. Conc t a) -> IO (Maybe a)

Run a concurrent computation with a given Scheduler and initial state, returning `Just result` if it terminates, and Nothing if a deadlock is detected.

Note how the t in Conc is universally quantified, what this means in practice is that you can't do something like this:

runConc (\s _ (x:_) -> (x, s)) () $ new >>= return

So CVars cannot leak out of the Conc computation. If this is making your head hurt, check out the "How runST works" section of https://ocharles.org.uk/blog/guest-posts/2014-12-18-rank-n-types.html

runConc' :: Scheduler s -> s -> (forall t. Conc t a) -> IO (Maybe a, s)

variant of runConc which returns the final state of the scheduler.

liftIO :: IO a -> Conc t a

Lift an IO action into the Conc monad.

Caution! Blocking on the action of another thread in liftIO cannot be detected! So if you perform some potentially blocking action in a liftIO the entire collection of threads may deadlock! You should therefore keep IO blocks small, and only perform blocking operations with the supplied primitives, insofar as possible.

spawn :: Conc t a -> Conc t (CVar t a)

Run the provided computation concurrently, returning the result.

fork :: Conc t () -> Conc t ()

Run the provided computation concurrently.

Communication: CVars

data CVar t a

The concurrent variable type used with the Conc monad. Internally, these are implemented as IORefs, but they are structured to behave fairly similarly to MVars. One notable difference is that MVars are single-wakeup, and wake up in a FIFO order. Writing to a CVar wakes up all threads blocked on reading it, and it is up to the scheduler which one runs next. Taking from a CVar behaves analogously.

newEmptyCVar :: Conc t (CVar t a)

Create a new empty CVar.

putCVar :: CVar t a -> a -> Conc t ()

Block on a CVar until it is empty, then write to it.

tryPutCVar :: CVar t a -> a -> Conc t Bool

Put a value into a CVar if there isn't one, without blocking.

readCVar :: CVar t a -> Conc t a

Block on a CVar until it is full, then read from it (without emptying).

takeCVar :: CVar t a -> Conc t a

Block on a CVar until it is full, then read from it (with emptying).

tryTakeCVar :: CVar t a -> Conc t (Maybe a)

Read a value from a CVar if there is one, without blocking.

Scheduling

type Scheduler s = s -> ThreadId -> [ThreadId] -> (ThreadId, s)

A Scheduler maintains some internal state, s, takes the ThreadId of the last thread scheduled, and the list of runnable threads (which will never be empty). It produces a ThreadId to schedule, and a new state.

Note: In order to prevent deadlock, the Conc runtime will assume that a deadlock situation has arisen if the scheduler attempts to (a) schedule a blocked thread, or (b) schedule a nonexistant thread. In either of those cases, the computation will be halted.

type ThreadId = Int

Every thread has a unique identitifer. These are implemented as integers, but you shouldn't assume they are necessarily contiguous.

randomSched :: RandomGen g => Scheduler g

A simple random scheduler which, at every step, picks a random thread to run.

randomSchedNP :: RandomGen g => Scheduler g

A random scheduler which doesn't pre-empt the running thread. That is, if the last thread scheduled is still runnable, run that, otherwise schedule randomly.

roundRobinSched :: Scheduler ()

A round-robin scheduler which, at every step, schedules the thread with the next ThreadId.

roundRobinSchedNP :: Scheduler ()

A round-robin scheduler which doesn't pre-empt the running thread.