Control.Monad.Conc.Fixed.Internal

Description

Concurrent monads with a fixed scheduler.

Documentation

class (Functor m, Functor n, Monad m, Monad n) => Fixed m n r a | m -> n, m -> r, m -> a, r -> m, a -> m where

A little more abstraction is required to be able to use different schedulers. We need to be parametrised by the specific action, reference, and underlying monad types used by our implementation.

Methods

newRef :: x -> n (r x)

Create a new reference

readRef :: r x -> n x

Read a reference.

writeRef :: r x -> x -> n ()

Overwrite the contents of a reference.

liftN :: n x -> m x

Lift an action from the underlying monad

stopA :: a

Action to stop

step :: a -> ThreadId -> Threads a -> n (Threads a, Maybe ThreadAction)

Evaluate a thread one step, by deconstructing the action and calling the appropriate step function from this module.

unC :: m x -> Cont a x

Unpack the continuation-based computation from its wrapping type.

type ThreadId = Int

Every thread has a unique identitifer. These are implemented as integers, but you shouldn't assume they are necessarily contiguous.

type Scheduler s = s -> ThreadId -> [ThreadId] -> (ThreadId, s)

A Scheduler maintains some internal state, s, takes the ThreadId of the last thread scheduled, and the list of runnable threads (which will never be empty). It produces a ThreadId to schedule, and a new state.

Note: In order to prevent deadlock, the Conc runtime will assume that a deadlock situation has arisen if the scheduler attempts to (a) schedule a blocked thread, or (b) schedule a nonexistant thread. In either of those cases, the computation will be halted.

type Trace = [(ThreadId, ThreadAction)]

One of the outputs of the runner is a Trace, which is just a log of threads and actions they have taken.

data ThreadAction

All the actions that a thread can perform.

Constructors

Fork ThreadId
Start a new thread.

Put [ThreadId]
Put into a CVar, possibly waking up some threads.

BlockedPut
Get blocked on a put.

TryPut Bool [ThreadId]
Try to put into a CVar, possibly waking up some threads.

Read
Read from a CVar.

BlockedRead
Get blocked on a read.

Take [ThreadId]
Take from a CVar, possibly waking up some threads.

BlockedTake
Get blocked on a take.

TryTake Bool [ThreadId]
try to take from a CVar, possibly waking up some threads.

Lift
Lift an action from the underlying monad.

runFixed :: Fixed m n r a => Scheduler s -> s -> m x -> n (Maybe x)

Run a concurrent computation with a given Scheduler and initial state, returning `Just result` if it terminates, and Nothing if a deadlock is detected.

runFixed' :: Fixed m n r a => Scheduler s -> s -> m x -> n (Maybe x, s, Trace)

Variant of runConc which returns the final state of the scheduler and an execution trace.

data Block

A Block is used to determine what sort of block a thread is experiencing.

Constructors

WaitFull ThreadId

WaitEmpty ThreadId

type Threads a = Map ThreadId (a, Bool)

Threads are represented as a tuple of (next action, is blocked).

type R r a = r (Maybe a, [Block])

CVars are represented as a reference containing a maybe value, and a list of things blocked on it.

runThreads :: Fixed m n r a => Trace -> ThreadId -> Scheduler s -> s -> Threads a -> r (Maybe x) -> n (s, Trace)

Run a collection of threads, until there are no threads left.

A thread is represented as a tuple of (next action, is blocked).

Note: this returns the trace in reverse order, because it's more efficient to prepend to a list than append. As this function isn't exposed to users of the library, this is just an internal gotcha to watch out for.

stepFork :: Fixed m n r a => a -> a -> ThreadId -> Threads a -> n (Threads a, Maybe ThreadAction)

stepPut :: Fixed m n r a => R r x -> x -> a -> ThreadId -> Threads a -> n (Threads a, Maybe ThreadAction)

stepTryPut :: Fixed m n r a => R r x -> x -> (Bool -> a) -> ThreadId -> Threads a -> n (Threads a, Maybe ThreadAction)

stepGet :: Fixed m n r a => R r x -> (x -> a) -> ThreadId -> Threads a -> n (Threads a, Maybe ThreadAction)

stepTake :: Fixed m n r a => R r x -> (x -> a) -> ThreadId -> Threads a -> n (Threads a, Maybe ThreadAction)

stepTryTake :: Fixed m n r a => R r x -> (Maybe x -> a) -> ThreadId -> Threads a -> n (Threads a, Maybe ThreadAction)

stepLift :: Fixed m n r a => n a -> ThreadId -> Threads a -> n (Threads a, Maybe ThreadAction)

stepStop :: Fixed m n r a => ThreadId -> Threads a -> n (Threads a, Maybe ThreadAction)

goto :: Fixed m n r a => a -> ThreadId -> Threads a -> Threads a

Replace the Action of a thread.

block :: Fixed m n r a => R r x -> (ThreadId -> Block) -> ThreadId -> Threads a -> n (Threads a)

Block a thread on a CVar.

launch :: Fixed m n r a => a -> Threads a -> (Threads a, ThreadId)

Start a thread with the next free ID.

kill :: Fixed m n r a => ThreadId -> Threads a -> Threads a

Kill a thread.

wake :: Fixed m n r a => R r x -> (ThreadId -> Block) -> Threads a -> n (Threads a, [ThreadId])

Wake every thread blocked on a CVar read/write.