Control.Monad.Conc.SCT.Shrink

Description

Functions for attempting to find maximally simple traces producing a given result.

Trace shrinking

shrink :: Eq a => (Maybe a, SCTTrace) -> (forall t. Conc t a) -> SCTTrace

Attempt to find a trace with a minimal number of pre-emptions that gives rise to the desired output.

shrink' :: (Maybe a -> Bool) -> SCTTrace -> (forall t. Conc t a) -> SCTTrace

Variant of shrink which takes a function to determine if the result is erroneous in the same way.

shrinkIO :: Eq a => (Maybe a, SCTTrace) -> (forall t. Conc t a) -> IO SCTTrace

Variant of shrink for IO. See usual caveats about IO.

shrinkIO' :: (Maybe a -> Bool) -> SCTTrace -> (forall t. Conc t a) -> IO SCTTrace

Variant of shrinkIO which takes a function to determine if the result is erroneous in the same way. See usual caveats about IO.

Utilities

If you wanted to implement your own shrinking logic, these are the building blocks.

candidates :: SCTTrace -> [([Decision], ThreadId)]

Generate all candidate trace prefixes from a trace. These are produced by attempting to drop one pre-emption. Furthermore, the ThreadId of the thread which performed the dropped pre-emption is also returned.

try :: (Maybe a -> Bool) -> [Decision] -> (forall t. Conc t a) -> [SCTTrace]

Try all traces with a given trace prefix and return those which satisfy the predicate.

tryIO :: (Maybe a -> Bool) -> [Decision] -> (forall t. Conc t a) -> IO [SCTTrace]

Variant of try for IO See usual caveats about IO.

essential :: ([Decision], ThreadId) -> [SCTTrace] -> Bool

Given a list of SCTTraces which follow on from a given prefix, determine if the removed pre-emption is "essential". That is, every SCTTrace starts with the prefix followed by a pre-emption to the given thread.

simplest :: [SCTTrace] -> Maybe SCTTrace

Return the simplest SCTTrace from a collection. Roughly, the one with the fewest pre-emptions. If the list is empty, return Nothing.