Control.Monad.Conc.SCT.Tests

Description

Useful functions for writing SCT test cases for Conc computations.

Test cases

data Result

The results of a test, including information on the number of cases checked, and number of total cases. Be careful if using the total number of cases, as that value may be very big, and (due to laziness) will actually force a lot more computation!.

Constructors

Result

Fields

_pass :: Bool

Whether the test passed or not.

_casesChecked :: Int

The number of cases checked.

_casesTotal :: Int

The total number of cases.

runTest :: Predicate a -> (forall t. Conc t a) -> Result

Run a test using the pre-emption bounding scheduler, with a bound of 2.

runTestIO :: Predicate a -> (forall t. Conc t a) -> IO Result

Variant of runTest using IO. See usual caveats about IO.

Predicates

type Predicate a = [Maybe a] -> Result

A Predicate is a function which collapses a list of results into a Result.

deadlocksNever :: Predicate a

Check that a computation never deadlocks.

deadlocksAlways :: Predicate a

Check that a computation always deadlocks.

alwaysSame :: Eq a => Predicate a

Check that the result of a computation is always the same. In particular this means either: (a) it always deadlocks, or (b) the result is always Just x, for some fixed x.

Utilities

pAnd :: Predicate a -> Predicate a -> Predicate a

Compose two predicates sequentially.

pNot :: Predicate a -> Predicate a

Invert the result of a predicate.

toPredicate :: (Maybe a -> Bool) -> [Maybe a] -> Result

Convert a boolean function to a Result-producing predicate.

takeWhile' :: (a -> Bool) -> [a] -> ([a], Bool, Int)

Variant of takeWhile that also includes a count of results returned and whether it traversed the entire list.