Control.Monad.Conc.Class

Description

The monad-conc package provides a class of monads which implement concurrency with (from the perspective of the programmer) nondeterministic interleaving.

The major difference between this and the Par family of monads (in monad-par) is that results may be implicitly dependent on interleaving, giving rise to data races between threads, whereas Par was explicitly designed to give rise to deterministic concurrency.

Documentation

class Monad m => ConcFuture future m | m -> future where

ConcFuture is the monad-conc alternative of ParFuture. It abstracts Conc monads which support futures. In itself, this is not enough to implement nondeterminism, however the class is provided to remove the NFData constraints imposed by ParFuture.

Methods

spawn :: m a -> m (future a)

Create a concurrent computation for the provided action, and return a future which can be used to query the result.

readCVar :: future a -> m a

Block until a value is present in the future, and then return it. This does not "remove" the value from the future, multiple gets are possible, unlike takeMVar for example.

class ConcFuture cvar m => ConcCVar cvar m | m -> cvar where

ConcCVar builds on futures by allowing CVars which threads can read from and write to, possibly multiple times. This is the key difference with the Par monads, where it is illegal to write multiple times to the same IVar, which removes the possibility of data races.

A minimal implementation consists of fork, newCVar, tryPutCVar, and tryTakeCVar. The default implementations of takeCVar and putCVar, however, are very inefficient, and should probably always be overridden to make use of implementation-specific blocking functionality.

Minimal complete definition

fork, newEmptyCVar, tryPutCVar, tryTakeCVar

Methods

fork :: m () -> m ()

Forks a computation to happen concurrently. Communication may happen over CVars.

newEmptyCVar :: m (cvar a)

Creates a new empty CVar.

putCVar :: cvar a -> a -> m ()

Put a value into a CVar. If there is already a value there, this will block until that value has been taken, at which point the value will be stored.

putCVar cvar a = tryPutCVar cvar a >>= \b -> unless b $ putCVar cvar a

tryPutCVar :: cvar a -> a -> m Bool

Attempt to put a value in a CVar, returning True (and filling the CVar) if there was nothing there, otherwise returning False

takeCVar :: cvar a -> m a

Take a value from a CVar. This "empties" the CVar, allowing a new value to be put in. This will block if there is no value in the CVar already, until one has been put.

takeCVar cvar = tryTakeCVar cvar >>= maybe (takeCVar cvar) return

tryTakeCVar :: cvar a -> m (Maybe a)

Attempt to take a value from a CVar, returning `Just value` (and emptying the CVar) if there was something there, otherwise returning Nothing.