Control.Monad.Conc.SCT.Tests

Description

Useful functions for writing SCT test cases for Conc computations.

Documentation

doTests

Arguments

:: Show a

=> Bool
Whether to print test passes.

-> [(String, Result a)]
The test cases

-> IO Bool

Run a collection of tests (with a pb of 2), printing results to stdout, and returning True iff all tests pass.

Test cases

data Result a

The results of a test, including information on the number of cases checked, and number of total cases. Be careful if using the total number of cases, as that value may be very big, and (due to laziness) will actually force a lot more computation!.

Constructors

Result

Fields

_pass :: Bool

Whether the test passed or not.

_casesChecked :: Int

The number of cases checked.

_casesTotal :: Int

The total number of cases.

_failures :: [(Maybe a, SCTTrace)]

The failed cases, if any.

runTest :: Predicate a -> (forall t. Conc t a) -> Result a

Run a test using the pre-emption bounding scheduler, with a bound of 2.

runTestIO :: Predicate a -> (forall t. Conc t a) -> IO (Result a)

Variant of runTest using IO. See usual caveats about IO.

runTest' :: Int -> Predicate a -> (forall t. Conc t a) -> Result a

Run a test using the pre-emption bounding scheduler.

runTestIO' :: Int -> Predicate a -> (forall t. Conc t a) -> IO (Result a)

Variant of runTest' using IO. See usual caveats about IO.

Predicates

type Predicate a = [(Maybe a, SCTTrace)] -> Result a

A Predicate is a function which collapses a list of results into a Result.

deadlocksNever :: Predicate a

Check that a computation never deadlocks.

deadlocksAlways :: Predicate a

Check that a computation always deadlocks.

deadlocksSometimes :: Predicate a

Check that a computation deadlocks at least once.

alwaysSame :: Eq a => Predicate a

Check that the result of a computation is always the same. In particular this means either: (a) it always deadlocks, or (b) the result is always Just x, for some fixed x.

alwaysTrue :: (Maybe a -> Bool) -> Predicate a

Check that the result of a unary boolean predicate is always true. An empty list of results counts as True.

alwaysTrue2 :: (Maybe a -> Maybe a -> Bool) -> Predicate a

Check that the result of a binary boolean predicate is always true between adjacent pairs of results. An empty list of results counts as True.

If the predicate fails, *both* (result,trace) tuples will be added to the failures list.

somewhereTrue :: (Maybe a -> Bool) -> Predicate a

Check that the result of a unary boolean predicate is true at least once. An empty list of results counts as False.

somewhereTrue2 :: (Maybe a -> Maybe a -> Bool) -> Predicate a

Check that the result of a binary boolean predicate is true between at least one adjacent pair of results. An empty list of results counts as False.

If the predicate fails, *both* (result,trace) tuples will be added to the failures list.

Utilities

pAnd :: Predicate a -> Predicate a -> Predicate a

Compose two predicates sequentially.

pNot :: Predicate a -> Predicate a

Invert the result of a predicate.

rForgetful :: Result a -> Result ()

Throw away the failures information in a Result (useful for storing them in a list).